removed unused components, updated css to most
