+++ src/components/Button/button.css
@@ -88,7 +88,6 @@
   text-shadow: 0 -1px 0 0 color(white a(50%));
   @mixin button var(--gray5), 3 {
     outline: none;
-
   }
 }
