+++ src/components/CardBody/card-body.css
@@ -1,3 +1,3 @@
 .card-body {
-  padding: 3rem;
+  margin: 3rem;
 }
