+++ src/components/Modal/modal.css
@@ -16,7 +16,7 @@
   left: 0;
   margin: 0;
   margin-bottom: 1rem;
-  padding: 2rem;
+  padding: 3rem;
   padding-top: 4.5rem;
   padding-bottom: 4.5rem;
   width: 100%;
